Reorganize stores in more natural structure, fix relative imports
