+++ src/stores/battery.js
