+++ src/stores/index.js
@@ -1,11 +1,11 @@
 export * from './configuration'
-export * from './batteryLevel'
+export * from './battery'
 export * from './navigationDrawer'
 export * from './viewMode'
 export * from './layout'
 export * from './ros'
 export * from './gstreamer'
-export * from './joystick'
-export * from './steering'
-export * from './keyboardSteering'
-export * from './joystickSteering'
+export * from './steering/joystick'
+export * from './steering/steering'
+export * from './steering/keyboardSteering'
+export * from './steering/joystickSteering'
