+++ src/stores/viewMode.js
@@ -1,17 +1,22 @@
 import { getCurrentInstance, ref } from 'vue'
 import { defineStore } from 'pinia'
 
+export const Modes = {
+    Normal: 'normal',
+    Edit: 'edit',
+}
+
 export const useViewModeStore = defineStore('viewMode', () => {
-    const mode = ref('normal')
+    const mode = ref(Modes.Normal)
 
     const $vuetify = getCurrentInstance().proxy.$vuetify
     function editMode() {
-        mode.value = 'edit'
+        mode.value = Modes.Edit
         $vuetify.theme.themes.light.primary =
             $vuetify.theme.themes.light.primaryEdit
     }
     function normalMode() {
-        mode.value = 'normal'
+        mode.value = Modes.Normal
         $vuetify.theme.themes.light.primary =
             $vuetify.theme.themes.light.primaryNormal
     }
